Time: 11 ms (76.77%), Space: 29 MB (65.81%) - LeetHub

+++ 3296-minimum-number-of-seconds-to-make-mountain-height-zero/3296-minimum-number-of-seconds-to-make-mountain-height-zero.cpp
@@ -29,7 +29,7 @@ public:
         long long totalHeight = 0; //height
         for(int t : workerTimes){
             //sum of all reducinh height of worker at midTime
-            totalHeight += (long long)(sqrt((2.0 * midTime) / t + 0.25) - 0.5);
+            totalHeight += (sqrt((2.0 * midTime) / t + 0.25) - 0.5);
  
 
             if(totalHeight>=mountainHeight) return true;
